% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.convert.365togreg.R
\name{cal.convert.365togreg}
\alias{cal.convert.365togreg}
\title{Convert gridded 365-days field to Gregorian calendar}
\usage{
cal.convert.365togreg(org.series, yearstart, yearend, fill.vector,
  data.type = 0)
}
\arguments{
\item{org.series}{Original time series (complete years only!).}

\item{yearstart}{First year of original 365-day time series.}

\item{yearend}{Last year of original 365-day time series.}

\item{fill.vector}{Vector indicating after which days in a
given year NA days should introduced; output of function
\code{\link{cal.indices.365togreg()}}. If
\code{data.type}=1: filename of RData file containing the
fill vector only.}

\item{data.type}{If 0: fill vector is provided as an
R object  (default). If 1: fill vector is provided in an
RData file containing the fill vector only
(filename is provided by \code{fill.matrix}).}
}
\value{
Vector (365/366 x nyears): original time series but in
Gregorian calendar (original series filled with NAs).
}
\description{
\code{cal.convert.365togreg} converts a timeseries of
a 365-days calendar (complete years) to Gregorian calendar by
filling in NAs.
}
\examples{
\dontrun{
# convert daily series of a 365-days calendar for the years
# 1980 to 2009 to a Gregorian calendar. NA indices provided as vector
# 'indices'
cal.convert.365togreg.grid(series,1980,2009,indices,0)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
