% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.indices.360togreg.R
\name{cal.indices.360togreg}
\alias{cal.indices.360togreg}
\title{Compute indices for 360-days to Gregorian calendar conversion}
\usage{
cal.indices.360togreg(yearend, outfile = NA)
}
\arguments{
\item{yearend}{Last year to consider (implicitly, first year is year 1).}

\item{outfile}{If provided, the resulting matrix will be written into
this RData file (optional).}
}
\value{
Matrix (1:yearend,6): indices (columns) for a given 360-days
year (row) after which to append NA days.
}
\description{
\code{cal.indices.360togreg} determines the day-of-year (DOY)
indices for each individual year of a 360-days calendar (for a given
period) at which to introduce NAs in order to transfer series to a full
Gregorian calendar.
}
\examples{
\dontrun{
# Calculate DOY indices after which to fill in NAs for all years
of the period 1 to 2200. No outfile written.
cal.indices.360togreg(2200)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
