% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.indices.365togreg.doytoabs.R
\name{cal.indices.365togreg.doytoabs}
\alias{cal.indices.365togreg.doytoabs}
\title{Compute absolute indices for 365-days to Gregorian calendar
conversion}
\usage{
cal.indices.365togreg.doytoabs(infile, yearstart, yearend, outfile = NA)
}
\arguments{
\item{infile}{Rdata-file contianing one matrix with DOY positions for each
year to fill in NAs (years = rows 1:2200, days = columns, 1:6); output of
function \code{\link{cal.indices.365togreg}}.}

\item{yearstart}{Start year of the desired period.}

\item{yearend}{End year of the desired period.}

\item{outfile}{If provided, the resulting vector will be written into this
RData file (optional).}
}
\value{
Vector (positions to fill in NAs): absolute indices for the period
(in Gregorian calendar) at which to fill NA days.
}
\description{
\code{cal.indices.365togreg.doytoabs} determines absolute
indices in a full Gregorian calendar (for a given period) at which to
introduce NA days corresponding to day-of-year (DOY) indices for individual
years with a 365-days calendar. Basis: Vector providing the NA DOY indices
for each individual year.
}
\examples{
\dontrun{
# Determine absolute indices to fill in NAs for the period
# 1981 to 2010. Matrix with indices for individual years
# provided by infile. No outfile written.
cal.indices.365togreg.doytoabs(infile,1981,2010)
}
}
\author{
Sven Kotlarski, Iris Feigenwinter (MeteoSwiss)
}
