% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.is.leapyear.R
\name{helper.is.leapyear}
\alias{helper.is.leapyear}
\title{Check if a given year is a leap year}
\usage{
helper.is.leapyear(year)
}
\arguments{
\item{year}{Year to check.}
}
\value{
Logical. TRUE/1: year is leap year. FALSE/0: year
is not a leap year.
}
\description{
Helper function: Check if a given year is a leap year.
}
\examples{
\dontrun{
# Check if year 2000 is a leap year
helper.is.leapyear(2000)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
