% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.mat.sort.R
\name{helper.mat.sort}
\alias{helper.mat.sort}
\title{Sort a matrix by a column}
\usage{
helper.mat.sort(mat, n)
}
\arguments{
\item{mat}{Matrix to be sorted.}

\item{n}{Index of column that defines sorting (increasing values).}
}
\value{
Sorted matrix.
}
\description{
Helper function: Sort a matrix by a certain column
(increasing values).
}
\examples{
\dontrun{
# Sort matrix mat by 2nd column
helper.mat.sort(mat,2)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
