% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.generate.R
\name{meta.generate}
\alias{meta.generate}
\title{Generate GCM-RCM meta information}
\usage{
meta.generate(filename)
}
\arguments{
\item{filename}{Name of meta information file to be created.}
}
\value{
List containing meta information (the same object that is
saved to filename)
}
\description{
\code{meta.generate} creates a meta information file
containing required GCM and RCM information for use within CH2018.
}
\examples{
\dontrun{
# Create meta information file metainfo_CH2018.RData
generate.meta('metainfo_CH2018.RData')
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
