% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.cdf.doy.R
\name{qm.cdf.doy}
\alias{qm.cdf.doy}
\title{Compute CDF for each day-of-year with moving window}
\usage{
qm.cdf.doy(series, yearbegin, yearend, doywindow, minq, maxq, incq, var,
  pr.wet)
}
\arguments{
\item{series}{Time series of modelled / observed data.}

\item{yearbegin}{First year of series.}

\item{yearend}{Last year of series.}

\item{doywindow}{Total width of moving window [days].}

\item{minq}{Minimum quantile [0..1].}

\item{maxq}{Maximum quantile [0..1].}

\item{incq}{Quantile increment (bin size).}

\item{var}{Variable name ('pr' for precip).}

\item{pr.wet}{Wet day threshold [mm/day].}
}
\value{
List of 3:
$cdf.matrix: CDF for each DOY (matrix[365,nquantiles]).
$f.wet: wet day frequency for each DOY (vector[365]); NA if not
precipitation.
$cdf.wet.vector: wet day CDF for each DOY
(matrix[365,101]; all percentiles); NA if not precipitation.
}
\description{
\code{qm.cdf.doy} computes the cumulative distribution
function (CDF) for each day-of-year (DOY) with a moving window
(+/- half window - 0.5). In case of precipitation also the wet day
frequency and wet day CDF are computed. The 29th February is considered
for all DOY ranges that include 1st March (DOY=60).
}
\examples{
\dontrun{
# Compute DOY-dependent CDF for a daily series ranging from 1981 to 2010 and
# for min and max quantile of 0.01 and 0.99, respectively. Quantile
# increment: 0.01. Variable: Precipitation. Wet day threshold: 0.1 mm/day.
qm.cdf.doy(time.series,1981,2010,91,0.01,0.99,0.01,'pr',0.1)
}
}
\author{
Sven Kotlarski (MeteoSwiss), Jan Rajczak (ETH Zurich)
}
