% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.corfun.bulk.R
\name{qm.corfun.bulk}
\alias{qm.corfun.bulk}
\title{Compute QM correction function for bulk series}
\usage{
qm.corfun.bulk(cdf.mod.bulk, cdf.obs.bulk)
}
\arguments{
\item{cdf.mod}{Modelled CDF for bulk series (vector[nquantiles])}

\item{cdf.obs}{Observed CDF for bulk series (vector[nquantiles])}
}
\value{
List of 3:
$correction: An additive correction value for each quantile of the bulk
series (vector[nquantiles]).
$cdfmod: The original modelled CDF vector.
$cdfobs: The original observed CDF vector.
}
\description{
\code{qm.corfun.bulk} computes the additive QM correction
function for each quantile of the bulk series; basically CDF mod - CDF obs.
}
\examples{
\dontrun{
# Compute QM correction function based on bulk modelled and observed CDFs
qm.corfun.bulk(cdf.mod,cdf.obs)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
