% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.corfun.doy.R
\name{qm.corfun.doy}
\alias{qm.corfun.doy}
\title{Compute QM correction function for each day-of-year}
\usage{
qm.corfun.doy(cdf.mod, cdf.obs)
}
\arguments{
\item{cdf.mod}{Modelled CDF for each DOY (matrix[365,nquantiles])}

\item{cdf.obs}{Observed CDF for each DOY (matrix[365,nquantiles])}
}
\value{
List of 3:
$correction: An additive correction value for each quantile and each DOY
(matrix[365,nquantiles])
$cdfmod: The original modelled CDF matrix.
$cdfobs: The original observed CDF matrix.
}
\description{
\code{qm.corfun.doy} computes the additive QM correction
function for each quantile and for each day-of-year (DOY); basically
CDF mod [doy] - CDF obs [doy].
}
\examples{
\dontrun{
# Compute QM correction function based on modelled and observed CDFs
# for each DOY
qm.corfun.doy(cdf.mod,cdf.obs)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
