% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.doystring.R
\name{qm.doystring}
\alias{qm.doystring}
\title{Construct daily vector of day-of-year for a given period}
\usage{
qm.doystring(yearbegin, yearend)
}
\arguments{
\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
Vector: Day-of-year time series.
}
\description{
\code{qm.doystring} constructs a daily vector with the
length of the entire period (defined by first and last year) specifying
the day-of-year (DOY; 1:365) of each day. A leap day (Feb 29) is given
the value 999.
}
\examples{
\dontrun{
# Construct DOY vector for the period 1981 to 2010
qm.doystring(1981,2010)
}
}
\author{
Jan Rajczak (ETH Zurich), Sven Kotlarski (MeteoSwiss)
}
