% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.doywindowgen.R
\name{qm.doywindowgen}
\alias{qm.doywindowgen}
\title{Construct matrix of all DOYs to be considered for a moving
window application}
\usage{
qm.doywindowgen(window.width)
}
\arguments{
\item{window.width}{Length of moving window (number of days). Should be
an uneven number.}
}
\value{
List of 3:
$half.width: Half window width minus 0.5.
$window.width: Total width of moving window.
$doymatrix: Matrix indicating the DOYs to account for depending on
the actual centered DOY (matrix[365,window.width]).
}
\description{
\code{qm.doywindowgen} constructs a matrix [365,window.width]
that indicates for each day-of-year (DOY) the DOYs to be considered in
moving window apoplications.
}
\examples{
\dontrun{
# Construct DOY-matrix for 91-day moving window
qm.doywindowgen(91)
}
}
\author{
Sven Kotlarski (MeteoSwiss), Jan Rajczak (ETH Zurich)
}
