% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.empqm.bulk.fa.R
\name{qm.empqm.bulk.fa}
\alias{qm.empqm.bulk.fa}
\title{The actual QM function for the bulk series
considering frequency adaptation (TAKE CARE: NOT YET FULLY TESTED!!)}
\usage{
qm.empqm.bulk.fa(series, cdf.mod.bulk, cdf.obs.bulk, cdf.obs.wet.bulk,
  cor.fun.bulk, fwet.obs, fwet.mod, pr.wet, method = c("linear", "bin"),
  var, output.diag = FALSE, minq, maxq, incq)
}
\arguments{
\item{series}{Time series of modelled data, e.g. transient (control+)
scenario series.}

\item{cdf.mod.bulk}{CDF of the modelled time series in the calibration
period, estimated with function \link{qm.cdf.bulk} (vector[nquantiles]).}

\item{cdf.obs.bulk}{CDF of the observed time series in the calibration
period, estimated with function \link{qm.cdf.bulk} (vector[nquantiles]).}

\item{cdf.obs.wet.bulk}{CDF of the observed time series in the calibration
period for wet days only, estimated with function \link{qm.cdf.bulk}
(vector[nquantiles]).}

\item{fwet.obs}{Observed wet day frequency (output of function
\link{qm.cdf.doy}).}

\item{fwet.mod}{Modelled wet day frequency output of function
\link{qm.cdf.doy}).}

\item{pr.wet}{Internal precipitation wet day threshold [mm/day].}

\item{method}{QM method: "binary" -> correction for closest quantile is
used, "linear" -> linear interpolation of correction function between lower
and upper quantile.}

\item{var}{Variable to correct.}

\item{output.diag}{Diagnostic QM console output (TRUE) or not (FALSE);
default: FALSE.}

\item{minq}{Minimum quantile for correction function [0.01 .. 0.99];
note: should not be 0 (for correct handling of extremes)!}

\item{maxq}{Maximum quantile for correction function [0.01 .. 0.99];
note: should not be 1 (for correct handling of extremes)!}

\item{incq}{Quantile increment for correction function (bin size).}

\item{corr.fun.bulk}{Additive correction function (cdf A - cdf B), obtained by
function \link{qm.corfun.bulk}.}
}
\value{
List of 3:
$qm.input.series: Input series to QM (input argument series).
$qm.corrected.series: Corrected series.
$quantile.index: A series of quantile indices wrt. simulated series in calibration period.
}
\description{
\code{qm.empqm.bulk.fa} is the actual QM function for
the bulk series. Equivalent to function \link{qm.empqm.doy.fa} but
for the bulk series, i.e. disregarding a day-of-year (DOY) dependence
of the correction function.
}
\examples{
\dontrun{
# Carry out bulk QM with frequency adaptation. Standard quantiles.
# Variable: precipitation. Observed wet day frequency of 0.9, modelled
# wet-day frequency of 0.95. Wet-day treshold of 0.1 mm/day. Linear QM,
# diagnostic terminal output.
qm.empqm.bulk.fa(series, cdf.modelled.bulk, cdf.observed.bulk, cdf.observed.wet.bulk, correction.function.bulk, 0.9, 0.95, 0.1, 'linear', 'pr', TRUE, 0.01, 0.00, 0.01)
}
}
\author{
Sven Kotlarski (MeteoSwiss), Jan Rajczak (ETH Zurich)
}
