% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.empqm.bulk.R
\name{qm.empqm.bulk}
\alias{qm.empqm.bulk}
\title{The actual QM function for the bulk series without
frequency adaptation (TAKE CARE: NOT YET FULLY TESTED!!)}
\usage{
qm.empqm.bulk(series, cdf.mod.bulk, cor.fun.bulk, method = c("linear",
  "bin"), var, output.diag = FALSE, minq, maxq, incq)
}
\arguments{
\item{series}{Time series of modelled data, e.g. transient (control+)
scenario series.}

\item{cdf.mod.bulk}{CDF of the modelled time series in the calibration
period, estimated with function \link{qm.cdf.bulk} (vector[nquantiles]).}

\item{method}{QM method: "binary" -> correction for closest quantile is
used, "linear" -> linear interpolation of correction function between lower
and upper quantile.}

\item{var}{Variable to correct.}

\item{output.diag}{Diagnostic QM console output (TRUE) or not (FALSE);
default: FALSE.}

\item{minq}{Minimum quantile for correction function [0.01 .. 0.99];
note: should not be 0 (for correct handling of extremes)!}

\item{maxq}{Maximum quantile for correction function [0.01 .. 0.99];
note: should not be 1 (for correct handling of extremes)!}

\item{incq}{Quantile increment for correction function (bin size).}

\item{corr.fun.bulk}{Additive correction function (cdf A - cdf B), obtained by
function \link{qm.corfun.bulk}.}
}
\value{
List of 3:
$qm.input.series: Input series to QM (input argument series).
$qm.corrected.series: Corrected series.
$quantile.index: A series of quantile indices wrt. simulated series in
calibration period.
}
\description{
\code{qm.empqm.bulk} is the actual QM function for the bulk
series. Equivalent to function \link{qm.empqm.doy} but for the bulk series,
i.e. disregarding a day-of-year (DOY) dependence of the correction function,
and without frequency adaptation.
}
\examples{
\dontrun{
# Carry out bulk QM without frequency adaptation. Standard quantiles.
# Variable: precipitation. Linear QM, diagnostic terminal output.
qm.empqm.bulk(series, cdf.modelled.bulk, correction.function.bulk, 'linear', 'pr', TRUE, 0.01, 0.00, 0.01)
}
}
\author{
Sven Kotlarski (MeteoSwiss), Jan Rajczak (ETH Zurich)
}
