% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm.empqm.doy.R
\name{qm.empqm.doy}
\alias{qm.empqm.doy}
\title{Short The actual QM function considering the DOY without
frequency adaptation}
\usage{
qm.empqm.doy(series, doy.series, cdf.mod, cor.fun, method = c("linear",
  "bin"), var, output.diag = FALSE, minq, maxq, incq)
}
\arguments{
\item{series}{Time series of modelled data, e.g. transient (control+)
scenario series.}

\item{doy.series}{Corresponding series of DOYs (output of function
\link{qm.doystring}).}

\item{cdf.mod}{CDF of the modelled time series in the calibration period,
estimated with function qm.cdf.doy(...) (matrix[365,nquantiles]).}

\item{method}{QM method: "binary" -> correction for closest quantile is
used, "linear" -> linear interpolation of correction function between lower
and upper quantile.}

\item{var}{Variable to correct.}

\item{output.diag}{Diagnostic QM console output (TRUE) or not (FALSE);
default: FALSE.}

\item{minq}{Minimum quantile for correction function [0.01 .. 0.99];
note: should not be 0 (for correct handling of extremes)!}

\item{maxq}{Maximum quantile for correction function [0.01 .. 0.99];
note: should not be 1 (for correct handling of extremes)!}

\item{incq}{Quantile increment for correction function (bin size).}

\item{corr.fun}{Additive correction function (cdf A - cdf B), obtained by
function \link{qm.corfun.doy(...)}.}
}
\value{
List of 3:
$qm.input.series: Input series to QM (input argument series).
$qm.corrected.series: Corrected series.
$quantile.index: A series of quantile indices wrt. simulated series in
calibration period.
}
\description{
\code{qm.empqm.doy} is the actual QM function considering
the day-of-year (DOY) and not considering frequency adaptation for
precipitation.
}
\details{
The function estimates the index (quantile) in which an observation
falls with respect to the calibration period. Based on this quantile (and
the DOY) a correction is applied, the final quantile map. For the 29th
February the same correction function as for 1st March (DOY=60) is applied.
Old and new extremes beyond the min and max percentile considered are
corrected according to the correction of the min and max percentile,
respectively (first and last quantile considered therefore have to be 0.01
and 0.99 -> check in function \link{qm.doqm}).
}
\examples{
\dontrun{
# Carry out DOY-dependent QM without frequency adaptation. Standard
# quantiles. Variable: precipitation. Linear QM, diagnostic terminal output.
qm.empqm.doy(series, doy.series, cdf.modelled, correction.function, 'linear', 'pr', TRUE, 0.01, 0.00, 0.01)
}
}
\author{
Jan Rajczak (ETH Zurich), Sven Kotlarski (MeteoSwiss)
}
