% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selbox.get.closest.gridcell.R
\name{selbox.get.closest.gridcell}
\alias{selbox.get.closest.gridcell}
\title{Determine closest climate model grid cell}
\usage{
selbox.get.closest.gridcell(lon.point, lat.point, file.lonlatgrid)
}
\arguments{
\item{lon.point}{Real-world longitude of target (decimal system).}

\item{lat.point}{Real-world latgitude of target (decimal system).}

\item{file.lonlatgrid}{Filename of Rdata file containing list (coords)
of grid lon and lat coordinates (matrices longrid and latgrid). Output
of function \link{selbox.nclonlat.to.rdata}.}
}
\value{
Two-element Vector containing the x-(lon) and the y-index (lat)
of the closest grid cell.
}
\description{
\code{selbox.get.closest.gridcell} determines the closest
climate model grid cell for a given lon/lat coordinate pair (target).
Criterion: smallest Euclidean distance in lon-lat space. Returns error
if target lies completely out of grid range (as defined by grid cell
centers).
}
\examples{
\dontrun{
# Find closest grid cell for coordinate longitude 10.38 and latitude 42.25
selbox.get.closest.gridcell(10.38,42.25,file.lonlatgrid)
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
