% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selbox.nclonlat.to.rdata.R
\name{selbox.nclonlat.to.rdata}
\alias{selbox.nclonlat.to.rdata}
\title{Create RData file with grid lattitudes and longitudes}
\usage{
selbox.nclonlat.to.rdata(infile, outfile, varlon, varlat)
}
\arguments{
\item{infile}{Netcdf filename (file has to contain real-word lon
and lat fields).}

\item{outfile}{RData filename for output.}

\item{varlon}{Variable name for real-world longitude in netcdf file.}

\item{varlat}{Variable name for real-world latitude in netcdf file.}
}
\value{
Return value of save statement (NULL).
}
\description{
\code{selbox.nclonlat.to.rdata} reads two fields (real-world
longitude and latitude) from a netcdf file and stores them in
an RData file. List name coords with two components:
$longrid: Matrix(nx,ny) containing real-world longitudes.
$latgrid: Matrix(nx,ny) containing real-world latitudes.
}
\examples{
\dontrun{
# Write netcdf longitudes and latitudes to Rdata file
selbox.nclonlat.to.rdata(file.netcdf,file.rdata,'longitude','latitude')
}
}
\author{
Sven Kotlarski (MeteoSwiss)
}
