% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.seasmean_fullday.R
\name{timstats.seasmean_fullday}
\alias{timstats.seasmean_fullday}
\title{Seasonal means

\code{timstats.seasmean_fullday(series,yearbegin,yearend)} computes time series
of seasonal means (mean annual cycle at seasonal resolution; last
December value is skipped) based on a daily time series covering full
years.}
\usage{
timstats.seasmean_fullday(series, yearbegin, yearend)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
List with 4 elements (one for each season) containing vectors of
seasonal means. Last DJF mean (December only) is skipped.
}
\description{
Seasonal means

\code{timstats.seasmean_fullday(series,yearbegin,yearend)} computes time series
of seasonal means (mean annual cycle at seasonal resolution; last
December value is skipped) based on a daily time series covering full
years.
}
\examples{
\dontrun{
# Compute seasonal means for a daily time series covering
# the years 2000 to 2005
timstats.seasmean_fullday(series,2000,2005)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
