% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.ydaymean_fullday.R
\name{timstats.ydaymean_fullday}
\alias{timstats.ydaymean_fullday}
\title{Multi-year daily means}
\usage{
timstats.ydaymean_fullday(series, yearbegin, yearend)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
Vector (366) of multi-year daily means.
}
\description{
\code{timstats.ydaymean_fullday(series,yearbegin,yearend)}
computes multi-year daily means (mean annual cycle at daily resolution)
based on a daily time series covering full years.
}
\examples{
\dontrun{
# Compute multi year daily means for a daily time series
# covering the years 2000 to 2005
timstats.ydaymean_fullday(series,2000,2005)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
