% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.ydaymean_fullday_mw.R
\name{timstats.ydaymean_fullday_mw}
\alias{timstats.ydaymean_fullday_mw}
\title{Multi-year daily means with moving window}
\usage{
timstats.ydaymean_fullday_mw(series, yearbegin, yearend, winsize)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}

\item{winsize}{Moving window length (uneven number of days).}
}
\value{
Vector (366) of multi-year daily means. 29th February is NA.
}
\description{
\code{timstats.ydaymean_fullday_mw(series,yearbegin,yearend,winsize)}
computes multi-year daily means (mean annual cycle at daily resolution) employing
a moving window around the center day and based on a daily time series covering full
years. The 29th February is discarded (NA in output series).
}
\examples{
\dontrun{
# Compute multi year daily means for employing a window of 91 days for
# a daily time series covering the years 2000 to 2005
timstats.ydaymean_fullday_mw(series,2000,2005,91)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
