% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.ydayquant_fullday.R
\name{timstats.ydayquant_fullday}
\alias{timstats.ydayquant_fullday}
\title{Multi-year daily quantiles

\code{timstats.ydayquant_fullday(series,yearbegin,yearend,quant)}
computes multi-year daily quantiles based on a daily time series covering
full years}
\usage{
timstats.ydayquant_fullday(series, yearbegin, yearend, quant)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}

\item{quant}{Quantile to compute [0,1].}
}
\value{
Vector (366) of multi-year daily quantiles.
}
\description{
Multi-year daily quantiles

\code{timstats.ydayquant_fullday(series,yearbegin,yearend,quant)}
computes multi-year daily quantiles based on a daily time series covering
full years
}
\examples{
\dontrun{
# Compute multi year daily 95th quantile for a
# daily time series covering the years 2000 to 2005
timstats.ydayquant_fullday(series,2000,2005,0.95)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
