% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.ydayquant_fullday_mw.R
\name{timstats.ydayquant_fullday_mw}
\alias{timstats.ydayquant_fullday_mw}
\title{Multi-year daily quantiles with moving window

\code{timstats.ydayquant_fullday_mw(series,yearbegin,yearend,quant,winsize)}
computes multi-year daily quantiles employing a moving window around
the center day and based on a daily time series covering full
years. Quantiles are computed by taking into account all daily
values within the moving window (no simple averaging of daily quantiles).}
\usage{
timstats.ydayquant_fullday_mw(series, yearbegin, yearend, quant, winsize)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}

\item{quant}{Quantile to compute [0,1].}

\item{winsize}{Moving window length (uneven number of days).}
}
\value{
Vector (366) of multi-year daily quantiles. 29th February is NA.
}
\description{
Multi-year daily quantiles with moving window

\code{timstats.ydayquant_fullday_mw(series,yearbegin,yearend,quant,winsize)}
computes multi-year daily quantiles employing a moving window around
the center day and based on a daily time series covering full
years. Quantiles are computed by taking into account all daily
values within the moving window (no simple averaging of daily quantiles).
}
\examples{
\dontrun{
# Compute multi year daily 95th quantile employng a moving window of
# 91 days for a daily time series covering the years 2000 to 2005
timstats.ydayquant_fullday_mw(series,2000,2005,0.95,91)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
