% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.yearmean_fullday.R
\name{timstats.yearmean_fullday}
\alias{timstats.yearmean_fullday}
\title{Annual means

\code{timstats.yearmean_fullday(series,yearbegin,yearend)}
computes a time series of annual means based on a daily time
series covering full years.}
\usage{
timstats.yearmean_fullday(series, yearbegin, yearend)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
Vector of annual means
}
\description{
Annual means

\code{timstats.yearmean_fullday(series,yearbegin,yearend)}
computes a time series of annual means based on a daily time
series covering full years.
}
\examples{
\dontrun{
# Compute annual means for a daily time series covering
# the years 2000 to 2005
timstats.yearmean_fullday(series,2000,2005)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
