% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.ymonmean_fullday.R
\name{timstats.ymonmean_fullday}
\alias{timstats.ymonmean_fullday}
\title{Multi-year monthly means}
\usage{
timstats.ymonmean_fullday(series, yearbegin, yearend)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
Vector (12) of multi-year monthly means
}
\description{
\code{timstats.ymonmean_fullday(series,yearbegin,yearend)}
computes multi-year monthly means (mean annual cycle at monthly resolution)
based on a daily time series covering full years.
}
\examples{
\dontrun{
# Compute multi year monthly means for a daily time series
# covering the years 2000 to 2005
timstats.ymonmean_fullday(series,2000,2005)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
