% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timstats.yseasmean_fullday.R
\name{timstats.yseasmean_fullday}
\alias{timstats.yseasmean_fullday}
\title{Multi-year seasonal means

\code{timstats.yseasmean_fullday(series,yearbegin,yearend)}
computes multi-year seasonal means (mean annual cycle at seasonal resolution)
based on a daily time series covering full years.}
\usage{
timstats.yseasmean_fullday(series, yearbegin, yearend)
}
\arguments{
\item{series}{Time series (vector) at daily resolution and for full years
(full Gregorian calendar).}

\item{yearbegin}{First year of time series.}

\item{yearend}{Last year of time series.}
}
\value{
Vector (4) of multi-year seasonal means
}
\description{
Multi-year seasonal means

\code{timstats.yseasmean_fullday(series,yearbegin,yearend)}
computes multi-year seasonal means (mean annual cycle at seasonal resolution)
based on a daily time series covering full years.
}
\examples{
\dontrun{
# Compute multi year seasonal means for a daily time series covering
# the years 2000 to 2005
timstats.yseasmean_fullday(series,2000,2005)
}

}
\author{
Sven Kotlarski (MeteoSwiss)
}
